Change total size of file accumulator to long from int to handle >2GB files

+++ javasrc/src/org/ccnx/ccn/utils/ccngetfile.java
@@ -1,7 +1,7 @@
 /**
  * A CCNx command line utility.
  *
- * Copyright (C) 2008, 2009 Palo Alto Research Center, Inc.
+ * Copyright (C) 2008, 2009, 2010 Palo Alto Research Center, Inc.
  *
  * This work is free software; you can redistribute it and/or modify it under
  * the terms of the GNU General Public License version 2 as published by the
@@ -127,7 +127,7 @@ public class ccngetfile {
 			byte [] buffer = new byte[readsize];
 			
 			int readcount = 0;
-			int readtotal = 0;
+			long readtotal = 0;
 			//while (!input.eof()) {
 			while ((readcount = input.read(buffer)) != -1){
 				//readcount = input.read(buffer);
